Add tracking script (#237)
We didn't have any way to see what features were being used in NEST.
This adds Matomo (Piwik) support and wires up tracking for the main nav.

You can configure the application by adjusting the site-id.js file in
the assets folder. This file can be mounted as a volume within the Docker
container which enables the ability to update the site id with an id that
is supplied by an external Matomo installation (e.g. one that is not
installed with the rest of Aerie).

Testing
---------

This is difficult to test when it is not in the context of Aerie. You
could start the app and then set a breakpoint in the analytics service
to verify that _paq.push is getting called.

Within Aerie you would be able to see what is actually being sent. To
do that, you will check out this branch into Aerie using the script. Then
you start it using docker-compose and the base, analytics, and nest configs.

Notes
------

Resolves https://cae-jira.jpl.nasa.gov/browse/MPSSEQ-810