Build Docker containers for projects in Jenkins (#14)
This commit adds a Jenkinsfile and supporting scripts which semi-
intelligently determine what projects were changed in a given commit,
and builds a Docker image for each of them.

It also consolidates all of the separate Docker Compose files into
several top-level Docker Compose files which are roughly specific to
environments.

* Add Jenkinsfile
* Update adaptation and plan service Docker files
* Check in Blackbird dependency to avoid future API issues
* Remove docker directory in favor of consolidated compose files
* Document installation procedure in docs/installation.md
* Provide supporting scripts for the Jenkins build
* Add a configuration for running Tyk without the dashboard

Testing
---------

Follow the instructions in docs/installation.md to get the service
running locally. It is possible to run all three configurations locally.