Add maven modules (#52)
This makes all the maven projects into modules so that idea will display them with all the fancy bells and whistles. It also makes the schema loading work without the file-install plugin.

To test it, run `git clean -dfX` to delete ALL untracked and ignored files in git. Then open it in idea and wait for it to set up the project.

It also works in VS code and from the command line.