Build reference adaptation from Merlin SDK (#60)
Build reference adaptation from Merlin SDK

The previous adaptations have been built on Blackbird, which was used in
the adaptation service to parse the adaptation. However, we needed to
create a reference adaptation built using the Merlin SDK. To do this the
Merlin SDK needed to be updated with the refelection capabilities that
Blackbird has (even though I no longer think reflection is the right way).

* Refactor the Merlin SDK to use Builder pattern for everything
* Remove Blackbird jar file and references in the adaptation service
* Update the Adaptation service to read Merlin SDK adaptations
* Create the first iteration of the Banananation adaptation

**More detail on the new Adaptation style:**

The main difference between Blackbird and the Merlin SDK approach is
that Blackbird uses reflection to parse Activity Types which are in
actual Java files. Merlin SDK uses the Builder pattern to construct each
of the parts of an adaptation which are then used as a library in the
adaptation service. Using the approach defined in this PR solves several
key problems:

1. Not being able to get parameters via reflection (at least as the
   activity type classes were defined in Merlin SDK)
3. Clipper's YAML based activity dictionary would cause problems,
   because the Activity Types only exist in memory, no files.
2. Designating which models to use at runtime

It has several benefits in addition to solving those problems:

- Requires less boilerplate than separate Class files
- May outperform reflection (https://stackoverflow.com/a/19563000)
- Is very DSL-like which might be easier to understand

Testing
----------

1. Build Merlin SDK, Banananation, and Adaptation service, in that order
2. Start the adaptation service
3. Upload the Banananation adaptation
4. Request the list of activities for the new adaptation

You should see a list of the activities defined in the adaptation and
there shouldn't be an errors.