Load an adaptation for a simulation (#157)
Adaptations loaded for simulation needed to be loaded within their own
container via AMQP. This is an intial approach to this problem. It works
by sending a POST request to the `/simulation` endpoint (of the API
gateway) with a `scheduleId` (currently unused) and an `adaptationid`.
A message is then broadcast over AMQP which indicates that a specific
adaptation should be loaded for simulation. This request is handled by an
adaptation runtime container which loads the simulation.

Testing
------

* Build the codebase with `mvn install` from the root
* Run the system with `docker-compose -f docker-compose-dev.yml up`
* POST the banananation adaptation to http://localhost:8081/adaptations/
* Make a GET request for activities to verify that the adaptation can load:
  `http://localhost:27182/api/adaptations/<ADAPTATION_ID>/activities`
* POST a (fake) `scheduleId` and the `adaptationId` from above to
  `http://localhost:8081/simulation/` to load the adaptation.
* Verify that the message `Successfully loaded adaptation banananation`
  has been logged in the terminal.

Notes
-------

Resolves https://jira.jpl.nasa.gov/browse/AERIE-70

* Upgrade adaptation-runtime, aerie-sdk, and simulation to Junit5