[AERIE-341] POST activity instances (#297)
* Implement POST /plans/:plan_id/activity_instances

* Clean up plan validation

This removes an unused validation method, and removes a validation that would
require clients to specify an empty list of activity instances for a plan. Now
it will sensibly default to an empty list when `null` is provided. A test case
for repository back-ends is added that ensures that even when a plan with a
`null` list of activities is stored, the repository always returns an list (only,
empty if it was null).