Fix Maven coordinates and add module-info.java to aerie-sdk
Several developers have discovered issues running `mvn test` from the
repository top-level. `mvn install` works fine, and `mvn test` from within
an individual component works fine. But `mvn test` at the top level fails
due to an internal JVM crash while Maven is executing the Surefire plugin.
Furthermore, this only occurs when dealing with the `adaptation` component.
All other components seem essentially unaffected.

This has been a significant hazard to one's sanity, among several other
seemingly unrelated build errors that have been plaguing us. The VM crash
causes Maven to provide singularly unhelpful errors (par for the course TBH).

One of the most bizarre features of these errors is that, even if one
redirects STDOUT and STDERR to a file, one of the crash messages *still*
displays visible in the terminal. On account of this, Evan Davis suggested
we instrument Maven itself with `strace`, a syscall tracing utility, to see
if we can identify where that error is coming from. With a little elbow
elbow grease, this gave us our first significant clue.

(Aside: Did you know that Maven ultimately calls the `write` syscall *one
character at a time* to display its output? I sure didn't! This made grepping
the strace output more complicated than it really needed to be. Use buffered
writers!)

The VM spawned by the Surefire plugin is failing during boot-up, because it
is unable to find a module required by the `adaptation` component:
`aerie.sdk`. This is quite strange -- one would certainly expect Java to
complain about the module's absence during compilation, no? Alas, Maven
is a riddle wrapped in a mystery inside an enigma.

We thought that, perhaps, the <scope>compile</scope> applied to the aerie-sdk
dependency might be a problem. If it's only available during compilation, and
not during the test phase, then that would explain its sudden absence. Yet
removing the scope restriction had no effect.

Later, we discovered that the `aerie.sdk` module is an *automatic* module,
meaning that its name is derived from the name of the JAR file it is packaged
as. On a hunch, we added a `module-info.java` and so gave it a formal name.
Et voila, everything works.

Yes, that's right.

Everything works.

Because we added a `module-info.java` file.

Because the module was otherwise an unnamed, automatic module.

Because Java is, frankly, nuts.

The module coordinate changes are just some cleanups identified in the course
of isolating this madness. It helps us keep all of our modules in a single
directory in ~/.m2/repository, so we can easily flush our cached packages when
debugging madness like this. It's not strictly necessary -- but in a world
with Maven in it, you need all the help you can get.