Remove the simulation component from the top-level pom
The simulation service exhibits a similar problem to the adaptation service's
insane `mvn test` VM crash. Thankfully, we're replacing this whole module
soon, and for now it isn't used at all. We'll just remove it from the pom
outright for now.

This also renames the Dockerfile because `scripts/build.sh` merely checks
if a changed component possesses a Dockerfile. If it does, it attempts to
build that Dockerfile. But since `simulation` isn't going to be built,
Jenkins will fail at this point.