[AERIE-428] Remove structural check on some bindings tests
I'm not really pleased about doing this, but our (de)serialization
situation is not great right now. On the other hand, in principle, the
the bindings tests should only be concerned with ensuring that the
right *kinds* of response entities are coming out of the bindings, not
that the entities themselves are actually correct. That responsibility
would better lie with tests on the (de)serialization facility. Only
the thinnest sliver of the system should really concern itself with
HTTP.