Remove `adaptationId` field from `Adaptation` class
It was only being used to construct `NoSuchActivityTypeException`s,
for which it shouldn't be necessary (since the lookup of an adaptation
by an ID is the domain of an entirely separate registry). A
higher-level module can receive this exception and transform it into
something containing both IDs if that is necessary for a higher
caller's scope.