Remove UnconstructableActivityInstanceException from App interface
The underlying UnconstructableActivityInstanceException is now
translated into a regular failure message in the normal list of
failures. This makes sense, as the entire purpose of the parameter
validation functionality is to verify that an instance can be
instantiated and is valid, and to report on failures of either
property.

The corresponding HTTP response for this functionality wraps the list
of failures in an object with a boolean "instantiable" key; if false,
a "failures" key will also be present containing the list of failures.