Look up adaptations in JAR by metadata
This is principally to fix an issue with running tests in IntelliJ
IDEA, where both banananation and sample-adaptation are on the
classpath. Sometimes, sample-adaptation is picked up when the tests
want banananation. This fix ensures that the correct adaptation is
always picked.

In practice, this also makes the adaptation loader more robust, as a
JAR may contain multiple adaptations and the correct adaptation will
still be chosen.