Simplify TreeLogger using tagless-final ideas
Information from "above" the present node should be passed in
as a parameter, while information from "below" the present node
can be part of the return value.

The information about whether to wrap a sub-expression in parenthese
is carried between the expression and its parent. Instead of introspecting
on the child from the parent, we now pass the parent information down
to the child. This de-duplicates some of the parenthesization logic.

See Oleg Kiselyov's copious notes on the topic
http://okmij.org/ftp/tagless-final/