Remove unnecessary Event type parameter from Query
The Event type parameter only showed up in the API of Query in the
getAt method, as part of the History<Scope, Event> being passed
in. But we don't actually need it -- we only use the database index
held by the history object, and the Scope parameter ensures that the
history we received is correlated with the database we're indexing.