Remove Scheduler#ask
Unlike the other methods on Scheduler, ask() has nothing to do with
the engine state. At the very least, it should be provided to the
adaptation in a different way.

Instead, we will allow the adaptation to own this logic, and provide
it back out to the driver by registering its solvers.

We will ensure that the driver can still provide symbolic
approximations to the front-end (for rendering, etc.) by requiring
that any solver be able to generate such symbolic approximations,
regardless of its internal type of dynamics.