Require adaptation to spawn an instantiated task spec
In the current type+arguments style of Scheduler#spawn and #defer,
the driver is responsible for looking up the adaptation's TaskSpecType
and instantiating it with the given arguments. This is a failable operation,
and dealing with it at the driver level raises uncomfortable questions,
like whether to abort or suspend the simulation, and how to inform the
task that invoked spawn/defer that the instantiation failed.

By accepting a pre-instantiated task spec, along with its governing
TaskSpecType, we force the adaptation to deal with these questions
on its own terms, and avoid polluting the engine with a responsibility
that properly rests with the adaptation in the first place.