Connect the adaptation service with the AdaptationFacade
Previously the ActivityMappers were used to deserialize and instantiate an activity
to retrieve schema, type, and parameter validation information. The adaptation
interface is now more clearly responsible for providing this information out and so
the AdaptationFacade no longer needs to de/serialize activities to retrieve this
information from the Adaptation.

Exception handling is changed so that the adaptation interface is not handling
exception outside its purview.

The getConstraintTypes call is returning an empty list and we are expecting it to be
implemented shortly.