Replace ProxyContext with DynamicCell
The purpose of ProxyContext is to provide a means of calling
contextual methods to Modules before those calls can actually be
serviced.  In other words, its role is to be a forward-reference to a
Context that will become available inthe future.

DynamicCell provides this same capability without delegating for the
underlying object's methods.  This might seem inconvenient, but the
only code that invokes Context methods directly is in the Module
superclass.  The delegation provided by ProxyContext has downsides:
whenever the Context interface changes, we have to update the rather
trivial implementations in ProxyContext.

Instead of having Module take a Context, it now takes a Supplier of a
Context, effectively separating these two concerns.  Since DynamicCell
implements Supplier, it acts as a general-purpose forward reference.