Set the active context within the ThreadedTask's thread
Because DynamicCell uses a ThreadLocal, it is important that a
ThreadedTask set the active context within the Thread, rather than
outside the Thread.  Otherwise, the DynamicCell won't actually appear
to be populated from the running task's perspective, as is the case
before this commit is applied.

It ends up simplest to use an inversion-of-control pattern to pass the
new Scheduler into the Thread after every yield.  We can also pass the
TaskStatus out of the Thread on every yield in the same way.