Rename Module to ModuleX
Since "Module" is also the name of a class in the java.lang package,
if our Module doesn't exist during annotation processing, the name is
instead eagerly resolved to the java.lang variant, causing compilation
to abort.

In order to avoid this, we'll temporarily rename ours to ModuleX,
pending a more cohesive review of Merlin's named concepts.