Replace waitUntil conditions with contextual getters
The main purpose of this commit is to remove the dependency from
Conditions to Solvers.  Over the next commits, any method within a
simulation context will become usable as a Resource, and this
necessitates setting up the context before querying the resource.

The Solver is currently the bridge between the driver and a resource,
so it will need to carry a Scoped<Context<?>>.  Since
CompoundConditions must specify a Solver, any Resource would need to
carry a Solver to be given to any condition derived from it.  This
isn't really a feasible requirement to put on a Resource.

The other purpose of this commit is to make conditions less "magic"
within the simulation framework. Instead of being special objects that
work with special solvers, they're just pieces of code that return a
duration (if any) specifying when the condition is next satisfied.
This parallels the aforementioned change to Resources.