Load an adaptation factory class by name
When running the DevAppDriver, the foo-adaptation is on the classpath
due to some tests that use it.  Unfortunately, ServiceLoader is very
sensitive to classpath changes, and when we request information about
a particular uploaded adaptation via the HTTP API, ServiceLoader often
picks up foo-adaptation instead of the JAR we meant to interrogate.

ServiceLoader makes sense to use when you don't know who is providing
a service.  In our case, we want to use a specific provider, not just
any provider we can find.

This commit changes our strategy to load a specific class by name.  We
identify the adaptation's factory class via the same mechanism as
ServiceLoader -- a file in the JAR's META-INF/services/ directory --
and then load that class directly via a classloader pointing to the
adaptation JAR.