Unset the daemon flag on most workers
Since daemons don't block the JVM from terminating, they can actually
mask resource leaks.  We should expect a clean termination when all is
well, but if we leak threads, the JVM should hang instead.  This might
seem like an inconvenience, but it's an important debugging signal.
We might never notice the resource leak otherwise!  Marking a thread
as a daemon means that any resource leaks caused by that thread or
threads under it will be ignored at termination, so we should only do
so in select circumstances.