Deprecate equals() on Register and similar models
We identified an issue in one of our modeled InSight activities where
Register#equals was being called on a Register<T> with a value of type
T. Despite the convenient name, the contract of an equals method is to
determine if two objects are semantically equivalent, not to determine
if an object contains a value. (In particular, if equals() is changed,
so too must be hashCode.) In the given situation, this constituted a
bug, because Register#equals would always return false.

Due to the high probability of semantic confusion, this commit marks
the Register#equals method, along with other methods on similar
classes, with the @Deprecated annotation, enabling the compiler to
help us identify locations where equals() was used incorrectly.