Remove Initializer#isBuilt
A mission model can determine whether it's in an initialization
context or not based on whether it's currently being called with an
Initializer, a Scheduler, or a Querier. In particular, we only care
about whether initialization has completed because of what that says
about what operations we're licensed to perform in the current
context.