Cancel the temporary cell in UncachedSimulationService
If the UncachedSimulationService is used with a simulation agent that
operates asynchronously, and hence expects a persistent location where
the results can be placed when they're ready, the cell will almost
always be Incomplete when the UncachedSimulationService returns. Since
the cell is a one-shot throwaway, we should at least cancel the cell
after getting the results so that the simulation agent knows it can
safely abort.