Simplify task breadcrumb tracking
Instead of caching the event history points at which a task performs a
query or action, we can cache the results of the queries (and the
number of actions). This leads to a much simpler approach where normal
operations are idempotized by a cache.

As a bonus (though I'd be lying if I didn't say this was the point of
this exercise), the Checkpoint class can go away entirely, simplifying
the Scheduler interface and its implementations.