Expose activity types from adaptation factory
Summary:

- Require activity types for adaptation building
- Replace `AdaptationFactory.Parameter` with `Parameter`
- Update `AdaptationFactory` interface to return task spec. types
- Update `AdaptationProcessor` generated source to return all activity types
- Update Merlin test extension to no longer rely on `ActivityTypes.register()`
- Replace `ResponseSerializers.serializeModelParameters()`
- Bubble up changes to simulation driver, adaptation service, profile builder
- Add `Phantom` to "brand" mission model instances

Explanation:

Change plugin interface to export actvity types separately from simulation.

Before, activity types were exported through a builder at initialization time.
With this change, they're provided by a method on the `AdaptationFactory` itself.
Since activities need a model instance to run against, the instantiate method
for the factory now returns a model instance, and the `TaskSpecType`
accepts this model instance when constructing a task.

Since activity types are no longer exported through the builder,
we also change how the Merlin JUnit extension is initialized,
so that it can be provided both a model instance and activity types
if a test is testing activities.