Run ThreadedTasks in a thread pool
Using a thread pool reduces the overheads involved in creating and
starting new threads for every task. In principle, the number of
threads created should scale with the degree of concurrency in the
system, i.e. the maximum number of in-progress tasks, rather than the
total number of tasks ever. For most plans, the former should be
dwarfed by the latter.

In practice, there's some wiggle room, as the default pool
configuration will destroy threads that haven't been used within a
certain amount of time, meaning the number of threads used may
increase if demand comes in waves.