Derive an iterative implementation of EventGraph#evaluate
This implementation was automatically derived from the recursive
implementation via a technique called "defunctionalizing the
continuation".  It's "automatic" in the sense that no creativity is
required; but I did have to follow the recipe manually, of course.

It's good to keep the recursive implementation around, both for
ensuring the correctness of the derived implementation relative to the
natural one, and so that future changes can be made to the recursive
implementation, and then the defunctionalized implementation can be
tossed out and regenerated from scratch. (It's quite hard sometimes to
make edits directly to an iterative implementation; they can be very
sensitive and require nonlocal reasoning.)