Use the 'JsonObjectBuilder' to append schema:JsonValue and order:int as the json data sent to postgres
When uploading a new model jar. The models activityType's parameter order isn't preserved anymore when we swapped to Hasura. This is because json isn't designed to preserve any order.

'ValueSchema' are converted into their respected 'JsonValue' which prevents the ability to include an order value. I reworked the json object to include the schema and order for each parameter This value will represent the order of the parameter listed in the ui.

AERIE-1541