Added `ValueSchemaOrderedMap` to `PreparedStatmets`
Issue: When uploading a new model jar. The models activityType's parameter order isn't preserved anymore when we swapped to Hasura. This is because json isn't designed to preserve any order.

This method is used to keep the order preserved by having a pairing of (order number,ValueSchema). This change will still work with the serialzeMap method. The 'seralizedOrderEntry' will parse the paring and add an 'order' entry in the json.

AERIE-1541