CreateActivityTypeAction builds a string to entry and hook up everything.
Issue: When uploading a new model jar. The models activityType's parameter order isn't preserved anymore when we swapped to Hasura. This is because json isn't designed to preserve any order.

A map of 'String' to entry(ordered, ValueSchema) is created and passed to `setValueSchemaOrderedMap` The server will parse out the order and ValueSchema and send a valid json to postgres and Hasura. The 'ActivityType' entries in the DB now contained an order number for each parameter. The Aerie-UI can now sort with this 'order' entry keeping the order that Activity Implementation builder intended.

AERIE-1541