Added Merlin specific environment variables to docker-compose.yml
I moved the variables defined in the json.config into the docker-compose.yml. The json.config is no longer needed and it will be removed.

Majority of the Aerie system configuration are located in the docker-compose.yml. At first a centralized database was proposed, but it makes more sense to have a centralize file that is easy to configure by the end-user. Merlin will need to have its configuration variables moved to this file as well.

[AERIE-1538]