Hook up Merlin with the environment variables defined in the docker-compose.yml
The config.json was removed and we don't need to parse this file anymore when creating the AppConfiguration. The code will now create the AppConfiguration with the environment variables. I also added fallback variables if one isn't defined.

Majority of the Aerie system configuration are located in the docker-compose.yml. At first a centralized database was proposed, but it makes more sense to have a centralize file that is easy to configure by the end-user. Merlin will need to have its configuration variables moved to this file as well.

[AERIE-1538]