Remove AppConfigurationJsonMapper and it's test.
Removed this class and test because we are not using this code anymore. There is no json configuration file to parse because everything was moved to the docker-compose.yml. Cleaning up the codebase of unused code.

Majority of the Aerie system configuration are located in the docker-compose.yml. At first a centralized database was proposed, but it makes more sense to have a centralize file that is easy to configure by the end-user. Merlin will need to have its configuration variables moved to this file as well.

[AERIE-1538]