Fix start_offset calculation in postProfileSegment
Previously, the code inserting profile segments assumed that
the "Duration"s in the profiles were provided to it were in fact
offsets from the previous segment.

This commit makes that code iterate backward from the end of the
simulation since the duration associated with each profile segment is
in fact its "extent" rather than an offset from the previous profile
segment.