Move simulation state to simulation_dataset table
Simulation state was initially placed in the dataset table due to a
lack of understanding of the motivation for the structure of the
database.  It is now clear that a dataset is more general than the
data for a simulation and simulation-specific traits do not belong
there. Instead, they belong in the simulation_dataset table.