Introduce ability to associate a plan with an existing dataset
Though at this time we are only allowing one plan to be associated
with a dataset, we will expand to allow multiple plans to share
datasets in the future. The database logic for handling this
automatically was a little tricky to figure out, so I don't want to
throw away the work I've done here concerning that. Technically this
commit is adding "dead" code, but I prefer to think of it as code that
simply has not yet been born. It will be soon :)