Move simulation_dataset delete trigger to after deletion
One of our constraints on simulation_dataset causes a
simulation_dataset row to be deleted whenever its corresponding
dataset is deleted. This behavior makes sense, as without a dataset, a
simulation can't be associated with it. However, we also have a
trigger to delete the dataset when a simulation_dataset row is deleted
that runs before the simulation_dataset row is removed. As a result,
the simulation_dataset trigger runs, deletes the dataset, and the
constraint on simulation_dataset deletes the simulation_dataset
row. Then after the trigger completes, the original delete fails,
since the row has already been deleted. This fixes that issue by
moving the trigger to happen after the simulation_dataset delete
operation.