Remove use of revisions in postgres actions
As the revision fields of our data model are used to track when
simulations are outdated, we do not need to manually check and pass
around revisions. Instead, when we check to see if a simulation run
for a plan exists, we merely check if a simulation_dataset row exists
with up-to-date revisions. Furthermore, the revision fields can be
(and now are) set automatically when creating a row of
simulation_dataset using current values. Furthermore, updating of a
simulation does not require revisions as it previously did since the
update operation can use the dataset ID.