Remove plan revision from ResultsCellRepository interface
Continuing our effort to rid Merlin of revisions, we remove the plan revision from the ResultsCellRepository methods. This also led to changing the deallocate method to taking a results cell. In fact, previously the deallocate method was not quite correct, and could've deallocated the wrong cells in certain scenarios. This is due to the fact that the deallocate method of the PostgresResultsCellRepository had to look up the dataset to delete. The InMemoryResultsCellRepository also had to be updated to use the InMemoryPlanRepository to determine plan revisions, as working in-memory it still makes sense to use revisions, as we are emulating the database in memory.