Remove ControlTask
ControlTask critically depends on the ability to get the task ID of
spawned tasks. Unfortunately, we need to remove this ability to
support future work (specifically, easily identifying which tasks to
deliver exceptions to for failed children).

ControlTask was only ever an optimization, to avoid polluting the
engine's task queue with tasks that wouldn't be reached for many
steps, so we can easily undo that optimization here and come back to
it later.