Remove waitFor
To continue supporting the important use-cases previously supported by
`waitFor`, we make `call` a primitive instead. This puts limits on
which activities can wait on other activities; in particular, only an
activity's direct parent may wait on it. Such a restriction will be
essential for propagating exceptions up the decomposition tree.