Provide all request data in RequestRecord
The RequestRecord is a record of a database table, and thus should
contain all columns of data. When originally written, I had unclear
boundaries between postgres data records and scheduler-server models,
and this helps make that boundary more clear. A follow-up commit will
remove the use of a RequestRecord in the PostgresResultsCell,
replacing with only the static data that should not change.