Remove an unnecessary use of TaskSpecType::getName
Technically, this just relocates the wart; it doesn't eliminate it.
It's a design smell in the first place that we need to ask the
TaskSpecType its own name. TaskSpecTypes are reflective descriptions
of task specifications; their "name" is a product of being registered
and distinguished amongst others.

Nonetheless, better a wart encapsulated in a single module than a wart
that spans components.