Added volume to the command expansion server and fixed issue using the wrong DB name
When the command dictionary is uploaded, there needs to be a location where we save the command type libs on the filesystem. The server has been updated to use this storage

There was a bug where the command expansion service wasn't pointed to the correct database. I also updated the server to use the correct name as a default.

[AERIE-1717]