Set merlin log level and output file in docker-compose
Javalin instantiates its logger statically in life-before-main, so the log level
and file cannot be set by merlin at runtime. And because JAVA_OPTS is already a
env variable, I can't interpolate MERLIN_LOGGING into it. So the configuration
has to go directly in JAVA_OPTS, which is terrible, but it works.