[FEAT] Adds a simulated_activity view
Simulated activities have been stored in the span field and embedded
some important relationships in a json blob. This adds a view
that implements the required joins across that json blob and joins
together the tables necessary to provide the full information about the
simulated activity from the associated tables.