Use synchronized for constraints-dsl-compiler
We had chosen not to use synchronized on the scheduling side, because
the compiler is invoked in a single-threaded, blocking manner. On the
constraints side, we are not currently queuing up constraints checking
code, so we need to resort to synchronized until we can determine a
more principled path forward.