add unit tests for applyWhen
create tests for the applyWhen() feature. Needed to create a new WindowsWrapperExpression, as we are going to refactor the Goals to take an Expression<Windows> - if a user doesn't want anything fancy and just wants to pass a Windows object to a goal then, it needs to be wrapped in such an expression that just yields itself.