Load Activity Types from Database
There are two use cases currently loading Activity Types in the merlin-server:
1. refreshActivityTypes
2. Generating Typescript for constraints checking

Refreshing activity types requires loading the mission model jar,
because it is this action that populates the database with the
activity types.

Fetching activity types must occur after refreshing the activity
types, and can fetch them from the database, reducing the need to load
the mission model jar.