Create a Unit Value Mapper
Add a Unit Value Mapper to map the UNIT value to an empty JAVA Map. This will be interpreted by the TS library generator as {}

-- Summary --
In the mission model, the activity's effect model that doesn't return a computed attribute (aka void run() method of the activity) uses UNIT which represents a non-null empty value. This is used internally by AERIE

https://en.wikipedia.org/wiki/Unit_type

This UNIT value will be exposed to our users when command authoring. Our users are not functional programmers so this causes confusion.

This will change the computed attribute to {} which makes more sense to the user.

[AERIE-2041]