Report strucutured activity instantiation failures
Previously "no such activity type" and "invalid arguments" activity
instantiation errors were just converted to a string and returned
through the Merlin server endpoint.

With the help of a sealed type we're now able to return the correct structure
within our response serializers depending on what the underlying error is.