Use property for publishing version string
The previous solution using `rootProject.ext.versionForPublishing` failed when building
Aerie locally from a separate Gradle project, likely due to `rootProject` taking on a different
value/meaning in that context.

Using a global property here we're able to specify the publishing version in the top-level build
and guarentee that any uses of it are localized to subprojects of our top-level build.