Migration Script
Here is a migration script. You can do a step-by-step migration on different databases or automatically apply/revert all available migrations.

It is a command line tool, so you can use python3 migration.py -h for information

If not run from the deployment directory, the path to the Hasura folder must be specified.

[AERIE-2118]