Define tree-structured plans that can be projected into a top-level Task
This representation allows us to schedule activities relative to other
activities, using the same API that tasks already use for delegation,
without duplicating that effort at the top-level engine API. It also
introduces the concept of a scheduling event, which is interpreted as
a task that merely does nothing for a period of time, but against
which activities can be scheduled relative to, ensuring that shifting
the event also shifts its dependents.

Representing such plans in the database is separate work!