Resolve typechecking issue by generating error in SeqJSON when C.<command>, A.<command>, and E.<command> are added to the immediate_command section of a sequence.
Currently, there is an issue with typechecking that permits the addition of C.<command>, A.<command>, and E.<command> to the immediate_command section of a sequence. Although this is invalid, TypeScript (TS) is unable to perform a type check due to the immediate type being a subset of the command type. To resolve this, a possible solution is to incorporate "type: immediate" to the Immediate type. However, the SeqJSON team does not want to include this redundancy in the JSON specification. Thus, as a workaround, an error is generated in SeqJSON to alert the user of this invalid action.