Fix RecurrenceGoal window vs Planning Horizon
This commit fix certain issues with the RecurrentGoal related to situations in which the recurrence of the goal or the windows in which the goal can be placed exceed the size of the plan horizon. To fix the issue, the plan horizon is added to the goal builder, so that getConflict method can make the corresponding checks. A number of tests have been modified to comply with this modification

Additional test for RecurrenceGoal
These tests check that different goal windows for RecurrenceGoal are properly scheduled