Add locking triggers to Tags Association Tables
activity_directive_tags was always supposed to be responsive to if the plan was locked

Additionally, if a tag is in use on a snapshot involved with an in-progress merge review, that tag cannot be deleted. This means that there is never a question of if a tag has been deleted mid-merge.