Close all running threads at the end of main
Without this, Javalin and the listener thread will continue running, even if the main loop has closed due to an exception. This prevents the container from exiting (meaning Docker won't restart it), and makes the worker appear alive but unresponsive to the notifications it is receiving when looking at the logs.