Add unit-awareness
Adds types and utilities for unit-awareness.

In particular, adds the `UnitAware` interface, which is generalized on the type of values to which units are attached.
This is to support adding units to both value types, like `Double`, and resource types, like `Resource<Polynomial>` or `Resource<Discrete<Double>>`.
Units are restricted to "absolute" units to simplify conversion, representation, and usage.
Units are represented by a Dimension and a floating-point scale compared to a base unit for that dimension.
Dimensions are represented exactly, as a product of rational powers of incommensurate base dimensions.

When using units on resources, we wrap UnitAware around Resource, and not vice-versa.
This applies a single unit to the entire lifetime of the resource, rather than letting the unit vary over time.
This lets us check dimensionality on resources once during initialization, then "bake in" constant conversion factors
to be used during simulation, a setup that has proven performant in Clipper's model.