Add more utilities for profiling resources.
Adds more support for profiling resource performance, especially around mutable resources.
Most importantly, it adds the static method `Resource.profileAllResources()`.
When called before constructing the model, this method turns on profiling for virtually every resource in the model.
This casts a broad net for early stages in a performance investigation, to highlight resources that are called unexpectedly frequently.