Pass expiry through LBCS and clampedIntegrate
Removes the use of eraseExpiry in LinearBoundaryConsistencySolver.
This was a patch put in place to support feedback loops involving the solver,
but it erases too much information. Leaving the expiry in place and erasing it
only when actually building a feedback loop lets downstream components like approximations
take that expiry into account.

Also reconfigures the use of eraseExpiry in clampedIntegrate to pass expiry information through.