Optimize clampedIntegrate
Rewrites PolynomialResources.clampedIntegrate to not use the LinearBoundaryConsistencySolver.

Profiling on the CADRE mission model suggests that clampedIntegrate was a primary bottleneck in time performance.
To address this, I'm applying the lessons learned from writing the LinearBoundaryConsistencySolver, mainly the importance of using `dominates` as the basis of polynomial comparisons, directly to clampedIntegrate.
Using this, I was able to rewrite the solver constraints as a non-iterative procedure, which queries the inputs exactly once each time the integral needs updating and limits the conditions on which we recompute the integral.