Add streamline Logger
Adds a Logger class to the streamline framework, which uses topics to emit log messages directly as events.
Also builds such a Logger when the Registrar is initialized, and reconfigures errors to go to that logger when the error behavior is set to LOG.
Before this, errors were collected awkwardly into a discrete string resource.