Move primary logger to Logging
Move the primary Logger instance to a separate class, out of Registrar.
This makes more sense logically, as Logging only incidentally depends on the Registrar, in the sense that creating the Registrar is a natural place to initialize all of the singletons.
It also makes the logger easier to find (I hope).