Add Multiplexing utilities
Adds a "Multiplexing" utility class implementing some patterns for multiplexing between different MutableResources.
These build "view" MutableResources, which don't have a cell of their own. Rather, they read values from and route effects to other MutableResources.

Also adds a "PrimenessModel" to the streamline-demo example, which demonstrates how multiplexing might be used to build "prime" and "backup"
views of A- and B-side resources, with the ability to emit effects on any of them and swap primeness on the fly.